refactor : 기능별 계층별 패캐지로 분류
기능별 계층별 패캐지로 분류
- 기능별 계층별로 패키지로 분류하여, 각 계층에 맞는 역할을 명확하게 하도록 수정
