+++ src/main/java/study/step3/RacingMain.java
@@ -1,5 +1,8 @@
 package study.step3;
 
+import study.step3.controller.CarRace;
+import study.step3.view.InputView;
+
 public class RacingMain {
   public static void main(String[] args) {
     int carNumber = InputView.inputCarNumber();
