+++ src/main/java/study/step3/domain/Car.java
@@ -1,4 +1,4 @@
-package study.step3;
+package study.step3.domain;
 
 import org.apache.commons.lang3.StringUtils;
