+++ src/main/java/study/step3/utils/MessageUtil.java
@@ -1,4 +1,4 @@
-package study.step3;
+package study.step3.utils;
 
 public class MessageUtil {
