+++ src/main/java/study/step3/validator/Validator.java
@@ -1,4 +1,6 @@
-package study.step3;
+package study.step3.validator;
+
+import study.step3.utils.MessageUtil;
 
 public class Validator {
